Move payment step further

+++ gen-image-bot/scripts/deploy-prod.sh
@@ -1,5 +1,13 @@
 HOST=image_bot@88.218.121.117
 
+read -p "Вы точно хотите продолжить, DEPLOYING ON PROD? (y/n): " confirm
+
+# Проверка ответа
+if [[ "$confirm" != "y" && "$confirm" != "Y" ]]; then
+    echo "Операция отменена пользователем."
+    exit 1
+fi
+
 ../gradlew build
 ssh $HOST 'mkdir -p ~/gen-image-bot/logs ~/gen-image-bot/trace'
 scp build/libs/gen-image-bot.jar "$HOST:~/gen-image-bot/gen-image-bot.jar"
